# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "XSHA Backend API service, providing user authentication, project management, Git credential management and other functions",
        "title": "XSHA Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/avatar/preview/{uuid}": {
            "get": {
                "description": "Preview avatar image by UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "image/*"
                ],
                "tags": [
                    "Admin Avatar"
                ],
                "summary": "Preview avatar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Avatar UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avatar image for preview",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid UUID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Avatar not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/avatar/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload an avatar image for administrator",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Avatar"
                ],
                "summary": "Upload avatar",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Avatar image to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avatar uploaded successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request parameter error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "File too large",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/avatar/{uuid}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update administrator's avatar by avatar UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Management"
                ],
                "summary": "Update admin avatar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Avatar UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Admin ID information",
                        "name": "adminData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "admin_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avatar updated successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request parameter error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Admin or avatar not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/login-logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get system login log records, supporting filtering by username, IP, success status, date range and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Get login logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username filter",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "IP address filter",
                        "name": "ip",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Success status filter",
                        "name": "success",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time filter (YYYY-MM-DD)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time filter (YYYY-MM-DD)",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number, defaults to 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size, defaults to 20, maximum 100",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login log list",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "logs": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                },
                                "page": {
                                    "type": "number"
                                },
                                "page_size": {
                                    "type": "number"
                                },
                                "total": {
                                    "type": "number"
                                },
                                "total_pages": {
                                    "type": "number"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get login logs",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/operation-logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get administrator operation log list with multi-condition filtering and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Logs"
                ],
                "summary": "Get operation log list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username filter",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Resource type filter",
                        "name": "resource",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Operation type filter",
                        "name": "operation",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Operation success status filter",
                        "name": "success",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time filter (YYYY-MM-DD)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time filter (YYYY-MM-DD)",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number, default is 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size, default is 20, maximum is 100",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation log list",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "logs": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                },
                                "page": {
                                    "type": "number"
                                },
                                "page_size": {
                                    "type": "number"
                                },
                                "total": {
                                    "type": "number"
                                },
                                "total_pages": {
                                    "type": "number"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get operation logs",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/operation-logs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information of a single operation log by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Logs"
                ],
                "summary": "Get operation log details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation log details",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "log": {
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid log ID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Log not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/operation-stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get operation statistics information within a specified time range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Logs"
                ],
                "summary": "Get operation statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username filter",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time (YYYY-MM-DD), default is 30 days ago",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time (YYYY-MM-DD), default is today",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation statistics information",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "end_time": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "operation_stats": {
                                    "type": "object"
                                },
                                "resource_stats": {
                                    "type": "object"
                                },
                                "start_time": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get operation statistics",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}